+++ radp-components/radp-commons/src/main/java/space/x9x/radp/commons/codec/binary/Base64.java
@@ -1,0 +1,11 @@
+package space.x9x.radp.commons.codec.binary;
+
+import lombok.experimental.UtilityClass;
+
+/**
+ * @author x9x
+ * @since 2024-12-27 16:25
+ */
+@UtilityClass
+public class Base64 extends org.apache.commons.codec.binary.Base64 {
+}