+++ radp-components/radp-commons/src/main/java/space/x9x/radp/commons/codec/digest/DigestUtils.java
@@ -1,0 +1,11 @@
+package space.x9x.radp.commons.codec.digest;
+
+import lombok.experimental.UtilityClass;
+
+/**
+ * @author x9x
+ * @since 2024-12-27 16:31
+ */
+@UtilityClass
+public class DigestUtils extends org.apache.commons.codec.digest.DigestUtils {
+}